Registration of user on imports

+++ src/main/java/br/com/alima/alurachallenge3/controllers/UserController.java
@@ -5,7 +5,6 @@ import br.com.alima.alurachallenge3.model.dto.NewUserDTO;
 import br.com.alima.alurachallenge3.model.dto.UpdateUserDTO;
 import br.com.alima.alurachallenge3.services.SystemUserService;
 import br.com.alima.alurachallenge3.services.exceptions.DataIntegrityViolationException;
-import br.com.alima.alurachallenge3.services.exceptions.ObjectNotFoundException;
 import lombok.AllArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Controller;
@@ -19,6 +18,7 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.validation.Valid;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Controller
 @RequestMapping("/users")
@@ -32,7 +32,7 @@ public class UserController {
     public String findAll(Model model) {
 
         List<SystemUser> users = userService.findAll();
-
+        users = users.stream().filter(u -> !u.getUsername().equals("admin@email.com.br")).collect(Collectors.toList()); //Remove user admin
         model.addAttribute("users", users);
         return "users";
     }
@@ -42,6 +42,8 @@ public class UserController {
 
         try {
             SystemUser user = userService.findById(id);
+            if (user.getUsername().equals("admin@email.com.br"))
+                throw new Exception("Prohibited action for this user.");
             mapper.map(user, dto);
             model.addAttribute("updateUserDTO", dto);
         } catch (Exception ex) {
@@ -56,6 +58,8 @@ public class UserController {
     public String resetPassword(@PathVariable Integer id, RedirectAttributes attributes, Model model) {
 
         try {
+            if (userService.findById(id).getUsername().equals("admin@email.com.br"))
+                throw new Exception("Prohibited action for this user.");
             userService.passwordReset(id);
         } catch (Exception ex) {
             attributes.addFlashAttribute("message", "User not registered.");
@@ -94,16 +98,22 @@ public class UserController {
         SystemUser user = mapper.map(dto, SystemUser.class);
 
         try {
+            if (userService.findById(user.getId()).getUsername().equals("admin@email.com.br"))
+                throw new Exception("Prohibited user!");
             userService.update(user);
-        } catch (DataIntegrityViolationException ex) {
-//            attributes.addFlashAttribute("message", "E-mail already registered.");
-            attributes.addFlashAttribute("message", ex.getMessage());
-            return "redirect:/users/"+dto.getId();
-        } catch (ObjectNotFoundException ex) {
-//            attributes.addFlashAttribute("message", "E-mail already registered.");
+        } catch (Exception ex) {
             attributes.addFlashAttribute("message", ex.getMessage());
             return "redirect:/users/"+dto.getId();
         }
+//        } catch (DataIntegrityViolationException ex) {
+////            attributes.addFlashAttribute("message", "E-mail already registered.");
+//            attributes.addFlashAttribute("message", ex.getMessage());
+//            return "redirect:/users/"+dto.getId();
+//        } catch (ObjectNotFoundException ex) {
+////            attributes.addFlashAttribute("message", "E-mail already registered.");
+//            attributes.addFlashAttribute("message", ex.getMessage());
+//            return "redirect:/users/"+dto.getId();
+//        }
 
         attributes.addFlashAttribute("message", "User edited successfully.");
         return "redirect:/users";
@@ -112,14 +122,20 @@ public class UserController {
     @GetMapping("/{id}/delete")
     public String delete(@PathVariable Integer id, NewUserDTO dto, RedirectAttributes attributes) {
         try {
+            if (userService.findById(id).getUsername().equals("admin@email.com.br"))
+                throw new Exception("Prohibited user!");
             userService.delete(id);
-        } catch (DataIntegrityViolationException ex) {
-            attributes.addFlashAttribute("message", ex.getMessage());
-            return "redirect:/users";
-        } catch (ObjectNotFoundException ex) {
+        } catch (Exception ex) {
             attributes.addFlashAttribute("message", ex.getMessage());
             return "redirect:/users";
         }
+//        } catch (DataIntegrityViolationException ex) {
+//            attributes.addFlashAttribute("message", ex.getMessage());
+//            return "redirect:/users";
+//        } catch (ObjectNotFoundException ex) {
+//            attributes.addFlashAttribute("message", ex.getMessage());
+//            return "redirect:/users";
+//        }
         attributes.addFlashAttribute("message", "User deleted successfully.");
         return "redirect:/users";
     }
